Added basic implementation of Enemy and BehaviorEnemy, removed code from AIStatePatrolComp
